fix(ci): keep the fleet rollup tests free of the supabase import chain
[#153]
